requirements.txt
certifi==2023.5.7
charset-normalizer==3.1.0
idna==3.4
instaloader==4.9.6
pip==23.1.2
python-dotenv==1.0.0
requests==2.31.0
setuptools==67.7.2
urllib3==2.0.2
wheel==0.40.0
